# Make CI faster