--- a/.travis.yml
+++ b/.travis.yml
@@ -32,8 +32,6 @@ addons:
       - python-dev
       - python3-dev
       - python3-pip
-      - liblua5.1-0-dev
-      - lua5.1
       - ruby-dev
 
 install: