--- a/.travis.yml
+++ b/.travis.yml
@@ -30,11 +30,9 @@ addons:
     packages:
       - language-pack-ja
       - vim
-      - libperl-dev
       - python-dev
       - python3-dev
       - python3-pip
-      - ruby-dev
 
 install:
   - rvm reset