--- a/scripts/coverage.sh
+++ b/scripts/coverage.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+
+set -ev
+
+if [[ "${THEMIS_PROFILE}" == "" ]]; then
+    exit
+fi
+
+covimerage write_coverage $THEMIS_PROFILE
+coverage xml
+bash <(curl -s https://codecov.io/bash)