--- a/scripts/install-vim.sh
+++ b/scripts/install-vim.sh
@@ -15,7 +15,10 @@ case "${TRAVIS_OS_NAME}" in
 	osx)
 		export HOMEBREW_NO_AUTO_UPDATE=1
 		brew update
-		brew install macvim --with-override-system-vim
+		brew install macvim
+		# Instead of --with-override-system-vim, manually link the executable because
+		# it prevents MacVim installation with a bottle.
+		ln -s "$(brew --prefix macvim)/bin/mvim" "/usr/local/bin/vim"
 		;;
 	*)
 		echo "Unknown value of \${TRAVIS_OS_NAME}: ${TRAVIS_OS_NAME}"