--- a/scripts/install-vim.sh
+++ b/scripts/install-vim.sh
@@ -16,8 +16,6 @@ case "${TRAVIS_OS_NAME}" in
 		make install
 		;;
 	osx)
-		export HOMEBREW_NO_AUTO_UPDATE=1
-		brew update
 		brew install macvim
 		# Instead of --with-override-system-vim, manually link the executable because
 		# it prevents MacVim installation with a bottle.