I made running CI faster.

Before: https://travis-ci.org/vim-jp/vital.vim/builds/390362502
After: https://travis-ci.org/rhysd/vital.vim/builds/390389523

Total: 5min 47sec -> 4min 2sec (1min 45sec faster)

Optimizations:

- Enabled a binary package (bottle) on macOS worker by removing installation options which prevented it. (4m54s -> 2m58s)
  - macOS worker is slow to start. It was a bottleneck.
  - Removed Lua dependency since Lua was not actually used in CI. (`--with-lua` could be omitted by this change)
  - Link `vim` executable directly (`--with-override-system-vim` could be omitted by this change)
- Separate running linters into another job.
  - Since it took 200s.
  - After optimizing macOS worker, the job which run linters was a bottleneck.

I assigned people as reviewers who would be familiar with
- CI scripts
- macOS

Fix #534