@ujihisa 2018-06-10
Great work! I'll leave proper review for someone who knows CIs better than me, but I like these changes :+1:

> Removed Lua dependency since Lua was not actually used in CI. (--with-lua could be omitted by this change)

Good catch! :+1: