@rhysd 2018-06-10
@ujihisa Thank you. I assigned you since you're more familiar with macOS. From macOS (Homebrew) user point of view, I'd be happy if you check the changes.