@ujihisa 2018-06-10
(I used to be a professional MacVim user. Currently I have no MacBooks/iMacs...)