@lambdalisue 2018-06-10
Note: It's a bit beyond this PR but I've found that using Docker is quite useful to test on Linux.

https://github.com/lambdalisue/gina.vim/blob/master/.travis.yml#L40-L43