@rhysd 2018-06-11
> Note: It's a bit beyond this PR but I've found that using Docker is quite useful to test on Linux.

Definitely. Moving to Docker container might be future work to improve this further.