@lambdalisue 2018-06-11
Outdated but

1. I think `--enable-luainterp` should be kept
2. I prefer `export THEMIS_VIM=$(brew ...)` instead