@rhysd 2018-06-11
@lambdalisue @raa0121 Tests use only `Text/Sexp.vim` (which has only one test case). Other modules which use Lua don't have any tests. So it's a trade-off. 'Give up one test case for `Text/Sexp.vim`' v.s. 'Making CI faster (about 1.4x faster). I don't have any strong opinion for this, actually.