@ujihisa 2018-06-11
Please ignore Text.Sexp. I'm planning to deprecate its lua version, along with https://github.com/vim-jp/vital.vim/pull/582 (Text.Sexp and brainf**k modules depended on it)