@tyru 2018-06-11
Still remaining issue exists https://github.com/vim-jp/vital.vim/pull/587#discussion_r194259022